:lipstick: style: header 파일 병합 및 style 파일 분리
