+++ client/src/App.tsx
@@ -3,8 +3,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 
 import ScrollToTop from "./utills/ScrollToTop";
-// import HeaderBefore from "./components/Header/HeaderBefore";
-import HeaderAfter from "./components/Header/HeaderAfter";
+import Header from "./components/Header/Header";
 import Footer from "./components/Footer/Footer";
 import Main from "./pages/Main/Main";
 import Signup from "./pages/Signup/Signup";
@@ -31,8 +30,7 @@ function App() {
     <>
       <BrowserRouter>
         <ScrollToTop />
-        {/* <HeaderBefore /> */}
-        <HeaderAfter popupHandler={popupHandler} />
+        <Header popupHandler={popupHandler} />
         <Routes>
           <Route index path="/" element={<Main detailHandler={detailHandler} setArticleId={setArticleId} />} />
           <Route index path="/signup" element={<Signup />} />
