+++ client/src/components/Header/style.tsx
@@ -1,23 +1,6 @@
-/*
-담당 : 송인선
-생성 : 2022.11.15
-수정 : 2022.11.16
-소개 : 헤더 컴포넌트 (로그인 전)
-설명 : 
-  - 페이지에서 공통적으로 사용되는 헤더 컴포넌트입니다.
-  - 현 단계에서는 타입 적용할 부분이 없었습니다.
-  - 반응형 적용
-  - 라우터 연결
-*/
-
 import styled, { keyframes } from "styled-components";
-import { ReactComponent as Logo } from "../../assets/img/logo.svg";
-import { ReactComponent as SearchIcon } from "../../assets/img/search-icon.svg";
-import { ReactComponent as MansaeCat } from "../../assets/img/mansae-cat.svg";
-import InnerContainer from "../InnerContainer/InnerContainer";
-import { Link } from "react-router-dom";
 
-const Header = styled.div`
+export const Header = styled.div`
   position: fixed;
   width: 100%;
   background-color: var(--color-white);
@@ -40,16 +23,7 @@ const FadeOut = (from: string, to: string) => keyframes`
   }
 `;
 
-const FadeIn = (from: string, to: string) => keyframes`
-  from {
-    transform: ${from};
-  }
-  to {
-    transform: ${to};
-  }
-`;
-
-const LogoBox = styled.div`
+export const LogoBox = styled.div`
   position: relative;
 
   @media screen and (max-width: 736px) {
@@ -59,10 +33,9 @@ const LogoBox = styled.div`
   .logo {
     width: 200px;
     height: 80px;
-    animation: ${FadeOut("0", "1")} 1s ease-in-out forwards;
 
     @media screen and (max-width: 736px) {
-      animation: ${FadeOut("1", "0")} 0.3s ease-in-out forwards;
+      opacity: 0;
     }
   }
 
@@ -72,15 +45,15 @@ const LogoBox = styled.div`
     left: 0px;
     width: 80px;
     height: 40px;
-    animation: ${FadeIn("scale(1)", "scale(0)")} 0.5s ease-in-out forwards;
+    opacity: 0;
 
     @media screen and (max-width: 736px) {
-      animation: ${FadeIn("scale(0)", "scale(1)")} 0.5s ease-in-out forwards;
+      opacity: 1;
     }
   }
 `;
 
-const SearchBox = styled.form`
+export const SearchBox = styled.form`
   margin: 0px 50px 0px;
   width: 80%;
   border: 2px solid var(--color-gray);
@@ -96,7 +69,7 @@ const SearchBox = styled.form`
   }
 `;
 
-const SearchInput = styled.input`
+export const SearchInput = styled.input`
   width: 60%;
   margin-left: 45px;
   height: 35px;
@@ -111,11 +84,24 @@ const SearchInput = styled.input`
   }
 `;
 
-const MenuBox = styled.div`
+export const MenuBox = styled.div`
+  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
   width: 20%;
+
+  .menu-icon {
+    position: absolute;
+    top: 0px;
+    right: 0px;
+    cursor: pointer;
+    display: none;
+
+    @media screen and (max-width: 736px) {
+      display: visible;
+    }
+  }
 `;
 
 const Small = (from: string, to: string, fontSmall: string, fontXSmall: string) => keyframes`
@@ -131,7 +117,28 @@ const Small = (from: string, to: string, fontSmall: string, fontXSmall: string) 
   }
 `;
 
-const Btn = styled.button`
+export const LoginBeforeBtn = styled.button`
+  margin: 0px 20px 0px 0px;
+  width: 40px;
+  height: 40px;
+  border-radius: 100%;
+  background-color: var(--color-white);
+  border: none;
+  flex-shrink: 0;
+  animation: ${FadeOut("0", "1")} 1s ease-in-out forwards;
+
+  @media screen and (max-width: 736px) {
+    animation: ${FadeOut("1", "0")} 0.3s ease-in-out forwards;
+  }
+
+  &:hover {
+    cursor: pointer;
+    transform: scale(1.1, 1.1);
+    transition: all 0.2s;
+  }
+`;
+
+export const LoginAfterBtn = styled.button`
   margin: 0px 16px 0px 0px;
   width: 60px;
   height: 60px;
@@ -144,6 +151,7 @@ const Btn = styled.button`
   animation: ${Small("45px", "60px", "var(--fs-pc-xsmall)", "var(--fs-pc-small)")} 0.5s ease-in-out forwards;
 
   @media screen and (max-width: 736px) {
+    display: none;
     animation: ${Small("60px", "45px", "var(--fs-pc-small)", "var(--fs-pc-xsmall)")} 0.5s ease-in-out forwards;
   }
 
@@ -154,34 +162,3 @@ const Btn = styled.button`
     transition: all 0.2s;
   }
 `;
-
-const HeaderBefore = () => {
-  return (
-    <Header>
-      <InnerContainer className="inner">
-        <LogoBox>
-          <Link to="/">
-            <Logo className="logo" />
-          </Link>
-          <Link to="/">
-            <MansaeCat className="logo-responsive" />
-          </Link>
-        </LogoBox>
-        <SearchBox>
-          <SearchIcon className="search-icon" />
-          <SearchInput type="text" placeholder="search..." />
-        </SearchBox>
-        <MenuBox>
-          <Link to="/login">
-            <Btn>Login</Btn>
-          </Link>
-          <Link to="/signup">
-            <Btn>Signup</Btn>
-          </Link>
-        </MenuBox>
-      </InnerContainer>
-    </Header>
-  );
-};
-
-export default HeaderBefore;
